+++ createComponents.js
@@ -1,62 +1,78 @@
-const fs = require('fs')
-const componentName = process.argv[2]
+const fs = require("fs");
 
-const componentTemplate = `import s from './${componentName}.module.css'
+const componentName = process.argv[2];
+const componentPath = process.argv[3];
+
+const componentTemplate = `import s from './${componentName}.module.scss';
 
 const ${componentName} = () => {
   return (
-    <div className={s.root}>${componentName}</div>
-  )
-}
+    <div className={s.root}>
+      
+    </div>
+  );
+};
 
-export default ${componentName}`;
+export default ${componentName};`;
 
 const indexTemplate = `import ${componentName} from './${componentName}';
 
-export default ${componentName};`
+export default ${componentName};`;
 
 const createComponents = new Promise((resolve, reject) => {
-    const path = `./src/components/${componentName}`
+  const path = `./src/${componentPath}/${componentName}`;
 
-    if (fs.existsSync(path)) {
-        reject('Component already exists')
+  if (fs.existsSync(path)) {
+    reject("Components is exist");
+  }
+
+  fs.mkdir(path, { recursive: true }, (err) => {
+    if (err !== null) {
+      reject(err);
+    } else {
+      resolve(path);
     }
-
-    fs.mkdir(path, { recursive: true }, (err) => {
-        if (err !== null) {
-            reject(err);
-        } else {
-            resolve(path);
-        }
-    });
+  });
 });
 
-console.log(`Start creating ${componentName} component`)
-
-createComponents.then((dirPath) => {
-    fs.writeFile(`${dirPath}/${componentName}.js`, componentTemplate, (err) => {
+console.log(`>>>>: Я начал создавать компонент ${componentName}`);
+createComponents
+  .then(async (dirPath) => {
+    await fs.writeFile(
+      `${dirPath}/${componentName}.js`,
+      componentTemplate,
+      (err) => {
         if (err !== null) {
-            Promise.reject(err)
+          Promise.reject(err);
         }
+      }
+    );
+
+    return dirPath;
+  })
+  .then(async (dirPath) => {
+    await fs.writeFile(
+      `${dirPath}/${componentName}.module.scss`,
+      ".root {\n\n}",
+      (err) => {
+        if (err !== null) {
+          Promise.reject(err);
+        }
+      }
+    );
+
+    return dirPath;
+  })
+  .then(async (dirPath) => {
+    await fs.writeFile(`${dirPath}/index.js`, indexTemplate, (err) => {
+      if (err !== null) {
+        console.log("Something wrong ", err);
+      }
     });
-    return dirPath
-}).then(async (dirPath) => {
-    fs.writeFile(`${dirPath}/${componentName}.module.css`, ".root {\n\n}", (err) => {
-        if (err !== null) {
-            Promise.reject(err)
-        }
-    })
-    return dirPath
-}).then(async (dirPath) => {
-    fs.writeFile(`${dirPath}/index.js`, indexTemplate, (err) => {
-        if (err !== null) {
-            Promise.reject(err)
-        }
-
-    })
-}).catch((e) => { console.log('We got a problem : ', e) })
-
-console.log('Creation copleted')
-
-
-
+    console.log(
+      `>>>>: Я закончил создавать компонент ${componentName}, его вы найдете по пути ${dirPath}`
+    );
+  })
+  .catch((err) => {
+    console.log("####: THIS IS ERROR - ", err);
+  });
