+++ src/pages/Main/Main.module.scss
@@ -1,0 +1,18 @@
+.cardSection {
+    display: flex;
+    justify-content: center;
+    padding: 24px 0;
+    background: #f7f7f7;
+}
+
+.cardTitle {
+    text-align: center;
+    margin-bottom: 32px;
+}
+
+.cardWrap {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 15px;
+    margin-bottom: 12px;
+}
